+++ src/Compare.php
@@ -19,6 +19,7 @@ namespace CloudCreativity\JsonApi\Testing;
 
 use Illuminate\Contracts\Routing\UrlRoutable;
 use Illuminate\Support\Arr;
+use Illuminate\Support\Collection;
 use SebastianBergmann\Comparator\ComparisonFailure;
 
 /**
@@ -86,12 +87,13 @@ class Compare
             return false;
         }
 
-        $members = collect($value);
+        $members = Collection::make($value);
 
         return $members->has('type') &&
             $members->has('id') &&
             !$members->has('attributes') &&
-            !$members->has('relationships');
+            !$members->has('relationships') &&
+            !$members->has('links');
     }
 
     /**
@@ -209,7 +211,7 @@ class Compare
             return [self::identifier($ids, $type)];
         }
 
-        return collect($ids)->map(function ($id) use ($type) {
+        return Collection::make($ids)->map(function ($id) use ($type) {
             return self::identifier($id, $type);
         })->values()->all();
     }
@@ -238,6 +240,10 @@ class Compare
 
         if (isset($id['id']) && $id['id'] instanceof UrlRoutable) {
             $id['id'] = (string) $id['id']->getRouteKey();
+        }
+
+        if (isset($id['id']) && is_int($id['id'])) {
+            $id['id'] = (string) $id['id'];
         }
 
         if ($type && !array_key_exists('type', $id)) {